Move attendance top bar to a separate component
